# Clean up / speed up ULEB128 decoding